move data formatter func setup into middleware utils
